Create README.md
This project will help you to follow the process of converting .py to OS_based script such .exe for windows and .sh for linux based operating system. as well as we will go through the flow for getting a basic system information on which this script will execute. 
I'll keep updating with each my research on it.